Rename treesitter config and remove un-needed pieces in wsl config
