+++ .config/nvim/lua/wsl-stuff.lua
@@ -16,11 +16,3 @@ vim.cmd[[
     \'coc-tsserver'
     \]
 ]]
-
-vim.g['coc_global_extensions'] = {'coc-go', 'coc-java', 'coc-json', 'coc-pairs', 'coc-pyright', 'coc-tsserver'}
--- autocmd User CocStatusChange redraws
-vim.api.nvim_create_autocmd("User", {
-    pattern = "CocStatusChange",
-    command = "redraws",
-    desc = "Fix no statusbar update without keypress"
-})
